requirements.txt
django-cms>=3.0